Split up tagger component
