+++ ui/v2.5/src/components/Tagger/utils.ts
@@ -1,5 +1,116 @@
 import * as GQL from "src/core/generated-graphql";
 import { getCountryByISO } from "src/utils/country";
+import { ParseMode } from "./constants";
+
+const months = [
+  "jan",
+  "feb",
+  "mar",
+  "apr",
+  "may",
+  "jun",
+  "jul",
+  "aug",
+  "sep",
+  "oct",
+  "nov",
+  "dec",
+];
+
+const ddmmyyRegex = /\.(\d\d)\.(\d\d)\.(\d\d)\./;
+const yyyymmddRegex = /(\d{4})[-.](\d{2})[-.](\d{2})/;
+const mmddyyRegex = /(\d{2})[-.](\d{2})[-.](\d{4})/;
+const ddMMyyRegex = new RegExp(
+  `(\\d{1,2}).(${months.join("|")})\\.?.(\\d{4})`,
+  "i"
+);
+const MMddyyRegex = new RegExp(
+  `(${months.join("|")})\\.?.(\\d{1,2}),?.(\\d{4})`,
+  "i"
+);
+const parseDate = (input: string): string => {
+  let output = input;
+  const ddmmyy = output.match(ddmmyyRegex);
+  if (ddmmyy) {
+    output = output.replace(
+      ddmmyy[0],
+      ` 20${ddmmyy[1]}-${ddmmyy[2]}-${ddmmyy[3]} `
+    );
+  }
+  const mmddyy = output.match(mmddyyRegex);
+  if (mmddyy) {
+    output = output.replace(
+      mmddyy[0],
+      ` ${mmddyy[1]}-${mmddyy[2]}-${mmddyy[3]} `
+    );
+  }
+  const ddMMyy = output.match(ddMMyyRegex);
+  if (ddMMyy) {
+    const month = (months.indexOf(ddMMyy[2].toLowerCase()) + 1)
+      .toString()
+      .padStart(2, "0");
+    output = output.replace(
+      ddMMyy[0],
+      ` ${ddMMyy[3]}-${month}-${ddMMyy[1].padStart(2, "0")} `
+    );
+  }
+  const MMddyy = output.match(MMddyyRegex);
+  if (MMddyy) {
+    const month = (months.indexOf(MMddyy[1].toLowerCase()) + 1)
+      .toString()
+      .padStart(2, "0");
+    output = output.replace(
+      MMddyy[0],
+      ` ${MMddyy[3]}-${month}-${MMddyy[2].padStart(2, "0")} `
+    );
+  }
+
+  const yyyymmdd = output.search(yyyymmddRegex);
+  if (yyyymmdd !== -1)
+    return (
+      output.slice(0, yyyymmdd).replace(/-/g, " ") +
+      output.slice(yyyymmdd, yyyymmdd + 10).replace(/\./g, "-") +
+      output.slice(yyyymmdd + 10).replace(/-/g, " ")
+    );
+  return output.replace(/-/g, " ");
+};
+
+export function prepareQueryString(
+  scene: Partial<GQL.SlimSceneDataFragment>,
+  paths: string[],
+  filename: string,
+  mode: ParseMode,
+  blacklist: string[]
+) {
+  if ((mode === "auto" && scene.date && scene.studio) || mode === "metadata") {
+    let str = [
+      scene.date,
+      scene.studio?.name ?? "",
+      (scene?.performers ?? []).map((p) => p.name).join(" "),
+      scene?.title ? scene.title.replace(/[^a-zA-Z0-9 ]+/g, "") : "",
+    ]
+      .filter((s) => s !== "")
+      .join(" ");
+    blacklist.forEach((b) => {
+      str = str.replace(new RegExp(b, "gi"), " ");
+    });
+    return str;
+  }
+  let s = "";
+
+  if (mode === "auto" || mode === "filename") {
+    s = filename;
+  } else if (mode === "path") {
+    s = [...paths, filename].join(" ");
+  } else {
+    s = paths[paths.length - 1];
+  }
+  blacklist.forEach((b) => {
+    s = s.replace(new RegExp(b, "gi"), " ");
+  });
+  s = parseDate(s);
+  return s.replace(/\./g, " ");
+}
 
 const toTitleCase = (phrase: string) => {
   return phrase
